Moved auth pages to own subfolder
