+++ src/screens/Auth/EmailVerify/EmailVerify.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .email-verify {
   max-width: 250px;
