+++ src/screens/Auth/Register/Register.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .register-form {
   margin: auto;
